Bugfix: Text meter not updating text to display
* Text meter wasn't updating the textToDisplay when the input audio was paused and dropped to negative infinity. Fixed conditional check in ValueHolder::updateHeldValue(float v) method.
Was checking if current value is greater than held value to determine when to update the text. But when the signal quickly dropped to negative infinity, held value updated (via the value holder timer callback?) before the textToDisplay had time to update. At which point, the updateHeldValue method would check if current value (-inf) was greater than heldValue (-inf), and would report false. Changed conditional to greater than or equal to, to cover this case.